add JS script for passing around ad_id URL parameter
The purpose of this is making sure that the parameter that will be added
by advertising agencies is passed along to `get.keycard.tech`. The issue
is that relative and absolute links to our own page need to be updated
too in order to retain the URL argument when someone visits `/docs/` for
example.

Signed-off-by: Jakub Sokołowski <jakub@status.im>